Add `Sync` bound to Send impl of LockWeak<T>`
It is possible to create data races by enclosing a
non-Sync type data to `ParentArc` and by accessing the enclosed data
from multiple threads via `ChildArc`.

This PR fixes the issue by adding a Sync bound to the
`Send` impl of `LockWeak<T>`.
For a detailed explanation with a proof-of-concept,
please refer to the corresponding PR description.

Thank you for reviewing :)